pyproject.toml
[tool.poetry]
name = "mader-project"
version = "0.1.0"
description = ""
authors = ["Leonardo Moura <leonardo.moura@eldorado.org.br>"]
readme = "README.md"
packages = [{include = "mader_project"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = {extras = ["standard"], version = ">=0.116.1,<0.117.0"}
pydantic-settings = ">=2.11.0,<3.0.0"
sqlalchemy = {extras = ["asyncio"], version = ">=2.0.43,<3.0.0"}
pwdlib = {extras = ["argon2"], version = ">=0.2.1,<0.3.0"}
tzdata = ">=2025.2,<2026.0"
pyjwt = ">=2.10.1,<3.0.0"
alembic = ">=1.16.5,<2.0.0"
psycopg = {extras = ["binary"], version = ">=3.2.10,<4.0.0"}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = ">=0.14.0,<0.15.0"
pytest = ">=8.4.2,<9.0.0"
pytest-cov = ">=7.0.0,<8.0.0"
taskipy = ">=1.14.1,<2.0.0"
factory-boy = ">=3.3.3,<4.0.0"
freezegun = ">=1.5.5,<2.0.0"
testcontainers = ">=4.13.2,<5.0.0"
 
[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev mader_project/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=mader_project -vv'
post_test = 'coverage html'
p = 'python'
ale = 'alembic'

[tool.coverage.run]
concurrency = ["thread", "greenlet"]